Extract Calendar to Cycles namespace
